+++ src/desktop/stb_impl.c
